+++ textblob/_text.py
@@ -941,25 +941,26 @@ class Sentiment(lazydict):
 # Unknown words are recognized as numbers if they contain only digits and -,.:/%$
 CD = re.compile(r"^[0-9\-\,\.\:\/\%\$]+$")
 
-def _suffix_rules(token, **kwargs):
+def _suffix_rules(token, tag="NN"):
     """ Default morphological tagging rules for English, based on word suffixes.
     """
-    word, pos = token
-    if word.endswith("ing"):
-        pos = "VBG"
-    if word.endswith("ly"):
-        pos = "RB"
-    if word.endswith("s") and not word.endswith(("is", "ous", "ss")):
-        pos = "NNS"
-    if word.endswith(("able", "al", "ful", "ible", "ient", "ish", "ive", "less", "tic", "ous")) or "-" in word:
-        pos = "JJ"
-    if word.endswith("ed"):
-        pos = "VBN"
-    if word.endswith(("ate", "ify", "ise", "ize")):
-        pos = "VBP"
-    return [word, pos]
+    if isinstance(token, (list, tuple)):
+        token, tag = token
+    if token.endswith("ing"):
+        tag = "VBG"
+    if token.endswith("ly"):
+        tag = "RB"
+    if token.endswith("s") and not token.endswith(("is", "ous", "ss")):
+        tag = "NNS"
+    if token.endswith(("able", "al", "ful", "ible", "ient", "ish", "ive", "less", "tic", "ous")) or "-" in token:
+        tag = "JJ"
+    if token.endswith("ed"):
+        tag = "VBN"
+    if token.endswith(("ate", "ify", "ise", "ize")):
+        tag = "VBP"
+    return [token, tag]
 
-def find_tags(tokens, lexicon={}, default=("NN", "NNP", "CD"), language="en", map=None, **kwargs):
+def find_tags(tokens, lexicon={}, model=None, morphology=None, context=None, entities=None, default=("NN", "NNP", "CD"), language="en", map=None, **kwargs):
     """ Returns a list of [token, tag]-items for the given list of tokens:
         ["The", "cat", "purs"] => [["The", "DT"], ["cat", "NN"], ["purs", "VB"]]
         Words are tagged using the given lexicon of (word, tag)-items.
@@ -968,34 +969,46 @@ def find_tags(tokens, lexicon={}, default=("NN", "NNP", "CD"), language="en", ma
         Unknown words that consist only of digits and punctuation marks are tagged CD.
         Unknown words are then improved with morphological rules.
         All words are improved with contextual rules.
+        If a model is given, uses model for unknown words instead of morphology and context.
         If map is a function, it is applied to each (token, tag) after applying all rules.
     """
     tagged = []
-    if isinstance(lexicon, Lexicon):
-        f = lexicon.morphology.apply
-    elif language == "en":
-        f = _suffix_rules
-    else:
-        f = lambda token, **kwargs: token
+    # Tag known words.
     for i, token in enumerate(tokens):
-        tagged.append([token, lexicon.get(token, i==0 and lexicon.get(token.lower()) or None)])
+        tagged.append([token, lexicon.get(token, i == 0 and lexicon.get(token.lower()) or None)])
+    # Tag unknown words.
     for i, (token, tag) in enumerate(tagged):
-        if tag is None:
-            if len(token) > 0 \
-            and token[0].isupper() \
-            and token[0].isalpha() \
-            and language != "de":
-                tagged[i] = [token, default[1]] # NNP
+        prev, next = (None, None), (None, None)
+        if i > 0:
+            prev = tagged[i-1]
+        if i < len(tagged) - 1:
+            next = tagged[i+1]
+        if tag is None or token in (model is not None and model.unknown or ()):
+            # Use language model (i.e., SLP).
+            if model is not None:
+                tagged[i] = model.apply([token, None], prev, next)
+            # Use NNP for capitalized words (except in German).
+            elif token.istitle() and language != "de":
+                tagged[i] = [token, default[1]]
+            # Use CD for digits and numbers.
             elif CD.match(token) is not None:
-                tagged[i] = [token, default[2]] # CD
+                tagged[i] = [token, default[2]]
+            # Use suffix rules (e.g., -ly = RB).
+            elif morphology is not None:
+                tagged[i] = morphology.apply([token, default[0]], prev, next)
+            # Use suffix rules (English default).
+            elif language == "en":
+                tagged[i] = _suffix_rules([token, default[0]])
+            # Use most frequent tag (NN).
             else:
-                tagged[i] = [token, default[0]] # NN
-                tagged[i] = f(tagged[i],
-                    previous = i > 0 and tagged[i-1] or (None, None),
-                        next = i < len(tagged)-1 and tagged[i+1] or (None, None))
-    if isinstance(lexicon, Lexicon):
-        tagged = lexicon.context.apply(tagged)
-        tagged = lexicon.entities.apply(tagged)
+                tagged[i] = [token, default[0]]
+    # Tag words by context.
+    if context is not None and model is None:
+        tagged = context.apply(tagged)
+    # Tag named entities.
+    if entities is not None:
+        tagged = entities.apply(tagged)
+    # Map tags with a custom function.
     if map is not None:
         tagged = [list(map(token, tag)) or [token, default[0]] for token, tag in tagged]
     return tagged